python/requirements.txt
web3
py-solc-x
dotenv